+++ internal/models/balance.go
@@ -2,6 +2,6 @@ package models
 
 type Balance struct {
 	//may be use float
-	Current   int `json:"current"`
-	Withdrawn int `json:"withdrawn"`
+	Current   int64 `json:"current"`
+	Withdrawn int64 `json:"withdrawn"`
 }
